chore: made calculate GCD task
